+++ server/routes/api/browser.py
@@ -29,11 +29,6 @@ bp = flask.Blueprint('api.browser', __name__, url_prefix='/api/browser')
 NO_MMETHOD_KEY = 'no_mmethod'
 NO_OBSPERIOD_KEY = 'no_obsPeriod'
 
-# Temporary fix for messy svgs. Remove once svgs have been fixed.
-BLOCKLISTED_STAT_VAR_GROUPS = {"dc/g/Establishment_Industry"}
-UPDATE_NUM_DESCENDENTS_SVG = {"dc/g/Establishment", "dc/g/Employment"}
-NUM_DESCENDENTS_TO_SUBTRACT = 12123
-
 
 @cache.memoize(timeout=3600 * 24)  # Cache for one day.
 @bp.route('/triples/<path:dcid>')
@@ -123,54 +118,6 @@ def get_observation_id():
         if mmethod == request_mmethod and obsPeriod == request_obsPeriod:
             result = dcid
             break
-    return Response(json.dumps(result), 200, mimetype='application/json')
-
-
-@bp.route('/statvar/search')
-@cache.cached(timeout=3600 * 24, query_string=True)
-def search_statvar():
-    """Gets the statvars and statvar groups that match the tokens in the query
-    """
-    query = request.args.get("query")
-    places = request.args.getlist("places")
-    result = dc.search_statvar(query, places,
-                               current_app.config["ENABLE_BLOCKLIST"])
-    return Response(json.dumps(result), 200, mimetype='application/json')
-
-
-@bp.route('/statvar/group')
-@cache.cached(timeout=3600 * 24, query_string=True)
-def get_statvar_group():
-    """Gets the stat var group node information.
-
-    This is to retrieve the adjacent nodes, including child stat vars, child stat
-    var groups and parent stat var groups for the given stat var group node.
-    """
-    stat_var_group = request.args.get("stat_var_group")
-    places = request.args.getlist("places")
-    result = dc.get_statvar_group(stat_var_group, places)
-    if current_app.config["ENABLE_BLOCKLIST"]:
-        childSVG = result.get("childStatVarGroups", [])
-        filteredChildSVG = []
-        for svg in childSVG:
-            svg_id = svg.get("id", "")
-            if svg_id in BLOCKLISTED_STAT_VAR_GROUPS:
-                continue
-            svg_num_descendents = svg.get("numDescendentStatVars", 0)
-            if svg_id in UPDATE_NUM_DESCENDENTS_SVG and svg_num_descendents > NUM_DESCENDENTS_TO_SUBTRACT:
-                svg["numDescendentStatVars"] = svg_num_descendents - NUM_DESCENDENTS_TO_SUBTRACT
-            filteredChildSVG.append(svg)
-        result["childStatVarGroups"] = filteredChildSVG
-    return Response(json.dumps(result), 200, mimetype='application/json')
-
-
-@bp.route('/statvar/path')
-@cache.cached(timeout=3600 * 24, query_string=True)
-def get_statvar_path():
-    """Gets the path of a stat var to the root of the hierarchy.
-    """
-    id = request.args.get("id")
-    result = dc.get_statvar_path(id)
     return Response(json.dumps(result), 200, mimetype='application/json')
